Add files via upload
Added RecipeCallAPI and WeatherAPICall.cs classes, which load data from openweathermap.com and from spoonacular. These classes are used in the homepage, which has been altered to have a button to demonstrate these api calls. 
I also changed the default page so that it hides the password, and has a little more information.
Right now, this will load a random recipe, and the current weather in Seattle.